Add files via upload
This is the Final.csv file that works for me, maybe there won't be index out of bounds error. If the error occurs, you will need to recopy the contents of the Final.csv into a new excel document, name it Final.csv and make sure to save it as .csv(comma-separated values).